SharedCore has more features now, which is cool, including a test to see if another object of the parent type has the same core, and another to clone the parent object.  That one is pretty cool, it means you can now get a real copy when you want to, great for multi-threaded stuff.
Also, two more classes are now SharedCore: Hash and Heap!

+++ src/archive.cpp
@@ -22,20 +22,15 @@ Bu::Archive::~Archive()
 {
 }
 
-void Bu::Archive::write( const void *pData, int32_t nSize )
+void Bu::Archive::write( const void *pData, size_t nSize )
 {
 	if( nSize == 0 || pData == NULL )
 		return;
 
-//	Bu::sio << "Writing starting at pos: " << rStream.tell() << " - "
-//		<< Bu::sio.flush;
-
 	rStream.write( (const char *)pData, nSize );
-//
-//	Bu::sio << rStream.tell() << " (" << nSize << "b)" << Bu::sio.nl;
 }
 
-void Bu::Archive::read( void *pData, int32_t nSize )
+void Bu::Archive::read( void *pData, size_t nSize )
 {
 	if( nSize == 0 || pData == NULL )
 		return;
